add a player scale option
this is entirely visual and doesn't modify the player's hitbox in any
capacity (at least, i hope so. f3+b says it doesn't, at least.)

this is really only *technically* implemented; the ui for this isn't
finished, and i don't quite expect that to change.

this was very much something i thought "how hard would it be to
implement", and it turns out the answer is "fairly easy".

i'm not really expecting this to go anywhere, and this in all likelihood
will just rot on this branch.

was still a fun idea for the 5 seconds of thought that i gave it though.